+++ test/unittests/libstorage/test_DagTransferPrecompiled.cpp
@@ -192,7 +192,7 @@ BOOST_AUTO_TEST_CASE(userAdd)
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
 
-    BOOST_TEST((256 - std::abs(CODE_INVALID_USER_NAME)) == result);
+    BOOST_TEST(CODE_INVALID_USER_NAME == result);
 
     // normal input, first add this user
     user = "user";
@@ -208,7 +208,7 @@ BOOST_AUTO_TEST_CASE(userAdd)
     params = abi.abiIn(userAddFunc, user, amount);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_USER_ALREADY_EXIST)) == result);
+    BOOST_TEST(CODE_INVALID_USER_ALREADY_EXIST == result);
 }
 
 BOOST_AUTO_TEST_CASE(userSave)
@@ -227,7 +227,7 @@ BOOST_AUTO_TEST_CASE(userSave)
     params = abi.abiIn(userSaveFunc, user, amount);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_USER_NAME)) == result);
+    BOOST_TEST(CODE_INVALID_USER_NAME == result);
 
     // invalid input, amount zero
     user = "user";
@@ -235,7 +235,7 @@ BOOST_AUTO_TEST_CASE(userSave)
     params = abi.abiIn(userSaveFunc, user, amount0);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_AMOUNT)) == result);
+    BOOST_TEST(CODE_INVALID_AMOUNT == result);
 
     // normal input, user is not exist, add this user.
     user = "user";
@@ -267,7 +267,7 @@ BOOST_AUTO_TEST_CASE(userSave)
     params = abi.abiIn(userSaveFunc, user, amount3);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_BALANCE_OVERFLOW)) == result);
+    BOOST_TEST(CODE_INVALID_BALANCE_OVERFLOW == result);
 }
 
 BOOST_AUTO_TEST_CASE(userDraw)
@@ -286,7 +286,7 @@ BOOST_AUTO_TEST_CASE(userDraw)
     params = abi.abiIn(userDrawFunc, user, amount0);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_USER_NAME)) == result);
+    BOOST_TEST(CODE_INVALID_USER_NAME == result);
 
     // invalid input, amount zero
     user = "user";
@@ -294,7 +294,7 @@ BOOST_AUTO_TEST_CASE(userDraw)
     params = abi.abiIn(userDrawFunc, user, amount1);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_AMOUNT)) == result);
+    BOOST_TEST(CODE_INVALID_AMOUNT == result);
 
     // add user first, after this the balance of "user" is 11111
     user = "user";
@@ -317,7 +317,7 @@ BOOST_AUTO_TEST_CASE(userDraw)
     params = abi.abiIn(userDrawFunc, user, amount4);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_INSUFFICIENT_BALANCE)) == result);
+    BOOST_TEST(CODE_INVALID_INSUFFICIENT_BALANCE == result);
 
     // get balance of this user
     dev::u256 balance;
@@ -345,7 +345,7 @@ BOOST_AUTO_TEST_CASE(userBalance)
     params = abi.abiIn(userBalanceFunc, user);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result, balance);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_USER_NAME)) == result);
+    BOOST_TEST(CODE_INVALID_USER_NAME == result);
 
     // normal input, user not exist
     user = "user";
@@ -353,7 +353,7 @@ BOOST_AUTO_TEST_CASE(userBalance)
     params = abi.abiIn(userBalanceFunc, user);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result, balance);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_USER_NOT_EXIST)) == result);
+    BOOST_TEST(CODE_INVALID_USER_NOT_EXIST == result);
 
     // create user
     user = "user";
@@ -377,7 +377,7 @@ BOOST_AUTO_TEST_CASE(userTransfer)
 
     std::string from, to;
     // dev::u256 amount;
-    dev::u256 result;
+    dev::s256 result;
     bytes out;
     bytes params;
 
@@ -386,7 +386,7 @@ BOOST_AUTO_TEST_CASE(userTransfer)
     params = abi.abiIn(userTransferFunc, from, to, amount0);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_USER_NAME)) == result);
+    BOOST_TEST(CODE_INVALID_USER_NAME == result);
 
     // invalid input, from user name empty string.
     from = "";
@@ -395,7 +395,7 @@ BOOST_AUTO_TEST_CASE(userTransfer)
     params = abi.abiIn(userTransferFunc, from, to, amount1);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_USER_NAME)) == result);
+    BOOST_TEST(CODE_INVALID_USER_NAME == result);
 
     // invalid input, to user name empty string.
     from = "from";
@@ -404,7 +404,7 @@ BOOST_AUTO_TEST_CASE(userTransfer)
     params = abi.abiIn(userTransferFunc, from, to, amount2);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_USER_NAME)) == result);
+    BOOST_TEST(CODE_INVALID_USER_NAME == result);
 
     // invalid input, amount zero.
     from = "from";
@@ -413,7 +413,7 @@ BOOST_AUTO_TEST_CASE(userTransfer)
     params = abi.abiIn(userTransferFunc, from, to, amount3);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_AMOUNT)) == result);
+    BOOST_TEST(CODE_INVALID_AMOUNT == result);
 
     // from and to user all not exist
     from = "from";
@@ -422,7 +422,7 @@ BOOST_AUTO_TEST_CASE(userTransfer)
     params = abi.abiIn(userTransferFunc, from, to, amount4);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_USER_NOT_EXIST)) == result);
+    BOOST_TEST(CODE_INVALID_USER_NOT_EXIST == result);
 
     // insert three user: user0(111111)  user1(2222222)
     // user3(0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff)
@@ -459,7 +459,7 @@ BOOST_AUTO_TEST_CASE(userTransfer)
     params = abi.abiIn(userTransferFunc, user0, user1, transfer);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_INSUFFICIENT_BALANCE)) == result);
+    BOOST_TEST(CODE_INVALID_INSUFFICIENT_BALANCE == result);
 
     dev::u256 balance;
     // get balance of user0
@@ -478,7 +478,7 @@ BOOST_AUTO_TEST_CASE(userTransfer)
     params = abi.abiIn(userTransferFunc, user1, user2, transfer);
     out = dtPrecompiled->call(context, bytesConstRef(&params), origin);
     abi.abiOut(bytesConstRef(&out), result);
-    BOOST_TEST((256 - std::abs(CODE_INVALID_BALANCE_OVERFLOW)) == result);
+    BOOST_TEST(CODE_INVALID_BALANCE_OVERFLOW == result);
 }
 
 BOOST_AUTO_TEST_SUITE_END()
