+++ test/unittests/libstorage/test_TableFactoryPrecompiled.cpp
@@ -76,15 +76,16 @@ BOOST_AUTO_TEST_CASE(call_afterBlock)
 {
     // createTable
     dev::eth::ContractABI abi;
-    bytes param =
-        abi.abiIn("createTable(string,string,string)", "t_test", "id", "item_name,item_id");
+    bytes param = abi.abiIn("createTable(string,string,string)", std::string("t_test"),
+        std::string("id"), std::string("item_name,item_id"));
     bytes out = tableFactoryPrecompiled->call(context, bytesConstRef(&param));
     Address addressOut;
     abi.abiOut(&out, addressOut);
     BOOST_TEST(addressOut == Address(0x1));
 
     // createTable exist
-    param = abi.abiIn("createTable(string,string,string)", "t_test", "id", "item_name,item_id");
+    param = abi.abiIn("createTable(string,string,string)", std::string("t_test"), std::string("id"),
+        std::string("item_name,item_id"));
     out = tableFactoryPrecompiled->call(context, bytesConstRef(&param));
     abi.abiOut(&out, addressOut);
     BOOST_TEST(addressOut <= Address(0x2));
@@ -92,7 +93,7 @@ BOOST_AUTO_TEST_CASE(call_afterBlock)
     // openTable not exist
     param.clear();
     out.clear();
-    param = abi.abiIn("openTable(string)", "t_poor");
+    param = abi.abiIn("openTable(string)", std::string("t_poor"));
     out = tableFactoryPrecompiled->call(context, bytesConstRef(&param));
     addressOut.clear();
     abi.abiOut(&out, addressOut);
@@ -100,7 +101,7 @@ BOOST_AUTO_TEST_CASE(call_afterBlock)
 
     param.clear();
     out.clear();
-    param = abi.abiIn("openTable(string)", "t_test");
+    param = abi.abiIn("openTable(string)", std::string("t_test"));
     out = tableFactoryPrecompiled->call(context, bytesConstRef(&param));
     addressOut.clear();
     abi.abiOut(&out, addressOut);
